server/requirements.txt
fastapi==0.99.1
mangum==0.17.0
python-dotenv
google-generativeai
requests